Allow .algorithm to go on a <div> wrapper
When using Markdown-style lists in Bikeshed, there's no way to add classes to the `<ol>` that gets generated. The usual practice in Bikeshed is to allow the class on an ancestor, so you can just use a wrapper `<div>`, but the styling here didn't allow that. My fix allows an immediate parent to have the .algorithm class instead.