Actually set the page background color
Currently the page background color is set on `body`, but further status-specific rules end up overriding that to transparent, so the specified color never gets used at all.  (And thus, can't be set differently by darkmode styles altering the `--bg` variable.)  I moved it to `html` instead, and moved the 'color' property as well, so they remain paired on the same element.

(Note that the `body` background is *not* generally lifted up to the canvas; some statuses apply the "UNOFFICIAL DRAFT" image via an `html` background, which overrides that. So setting bg color on the `html` is slightly better anyway, at least from a theoretical standpoint.)