Add meaningful text to toc controls
The current implementation does not have a label for navigation
elements, since the text is hidden. Firefox DevTools also claims
that the node does not have any label.

Instead, we can use `aria-labelledby` to reference to the hidden
element, to override the default text and have proper announcement.

Note that the code uses `const` in this place where others are
using `var`. However, at the bottom of this script, we already
use `const`, so I figured that the usage is fine.

Fixes #136